Added Repostories

+++ src/main/java/es/upm/etsit/dat/identi/persistence/repository/CensusMemberRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import es.upm.etsit.dat.identi.persistence.model.CensusMember;
 
-public interface CensusMemberRepository extends JpaRepository<CensusMember, Integer> {
+public interface CensusMemberRepository extends JpaRepository<CensusMember, Long> {
     CensusMember findByEmail(String email);
 
 }
